+++ src/Celavi/Exception/DomainException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Celavi\Exception;
+
+class DomainException extends \DomainException implements ExceptionInterface
+{
+}