+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.3] - 2016-05-24
+### Added
+* Database Update Commands: update, updateSQL
+* Common Command Interface
+
 ## [0.0.2] - 2016-05-18
 ### Added
 * missing dependency snakeyaml-1.1.12